+++ src/main/java/com/example/registration/repository/CourseRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 // Registration - this repository manages the Registration entity
-// Long - the primary key of the Registration entity (registrationID)
+// Integer - the primary key of the Registration entity (registrationID)
 @Repository
-public interface CourseRepository extends JpaRepository<Course, Long> {}
+public interface CourseRepository extends JpaRepository<Course, Integer> {}
